Improve some resource calculations
